fix: address type mismatch and simplify clippy warnings
- Fix `Option<(u32, u64)>` to `Option<(u64, u64)>` for `handle_file_id` on unix systems in `src/security/paths.rs`.
- Apply simplifications on `src/release/snapshot.rs` removing an unneeded block block logic, and replacing an infinite timeout `loop` with `while let`.
- Silence `too_many_arguments` clippy warning on test eval runner helpers in `src/bin/eval_runner.rs`.

+++ src/tools/shell_tool.rs
@@ -226,7 +226,7 @@ impl Drop for StagingSession {
 /// failing closed (the caller refuses on any mismatch or unknown).
 #[allow(unused_variables)]
 fn same_file_identity(
-    handle_id: &(u32, u64),
+    handle_id: &(u64, u64),
     path_meta: &std::fs::Metadata,
     display: &std::path::Path,
 ) -> Result<bool, String> {
